made transition from 32 bit potrace words (Integer) to 64 bit potrace word with LONG, and changed every method in potrace_bitmap to non static method

+++ src/Tools/BitmapPrinter.java
@@ -37,7 +37,7 @@ public class BitmapPrinter {
     }
 
     private String getCurrentPotraceWord(int currentLine, int currentPotraceWord) {
-        return Integer.toBinaryString(bitmap.map[currentLine*bitmap.dy + currentPotraceWord]);
+        return Long.toBinaryString(bitmap.map[currentLine*bitmap.dy + currentPotraceWord]);
     }
 
     private void fillUpPotraceWordWithZeros(String currentPotraceWord) {
